set it up to merge
