+++ src/main/java/frc/robot/commands/auto_commands/AutoRight.java
@@ -5,7 +5,7 @@
 // /* the project.                                                               */
 // /*----------------------------------------------------------------------------*/
 
-package frc.robot.commands;
+package frc.robot.commands.auto_commands;
 
 import edu.wpi.first.wpilibj2.command.CommandBase;
 import frc.robot.Robot;
